requirements-dev.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.5